Added support for lib64 site packages
Fixed issue with lib64 site packages not being included in lambda
package.  Lib64 packages will not be copied if the lib64 directory is
determined to be a symlink.  This commit also removes any use of the
shutil copytree method in favor of the copy_tree method in the local
utils package.