Simplified Package class and cleaned up the flow
Moved requirements and virtualenv to their own methods that handle
different scenarios based on what information is passed to them. If they
are not explicitly set then the default actions are performed. Moved a
lot of the methods to be private and meant to be called by the public
methods. Added a build() method to the Package class that can be called
to perform all major tasks needed to build a package.  Moving the
package validation to the uploader.