Change default config file location to "function directory/lambda.json"
Previously, without providing config file with -c argument, lambda-uploader would look for the config file in the directory from where it was executed.

This behaviour is not clearly stated in the README. The example also stores the config file and function in the same directory. Calling lambda-uploader from the root directory - `lambda-uploader example` - will raise an exception "lambda.json not a valid configuration file". 

This commit modifies the default config location to "function directory/lambda.json" to follow expected behaviour.